make routing table grow as nodes as inserted

+++ src/util.rs
@@ -5,14 +5,14 @@ pub const U64_HEX_NUM_OF_DIGITS: u32 = u64::BITS / HEX_NUM_OF_BITS_PER_DIGIT as 
 pub const HEX_DIGIT_MASK: u32 = 0xF;
 
 /// Gets the nth hexadecimal digit from a u64.
-pub fn get_nth_digit_in_u64_hex(num: u64, n: u32) -> Result<u32> {
+pub fn get_nth_digit_in_u64_hex(num: u64, n: usize) -> Result<u32> {
     if n >= 16 {
         return Err(crate::error::Error::Internal(
             "There are only 16 digits in a hexadecimal number.".into(),
         ));
     }
 
-    Ok((num >> (u64::BITS - (n + 1) * HEX_NUM_OF_BITS_PER_DIGIT)) as u32 & HEX_DIGIT_MASK)
+    Ok((num >> (u64::BITS - (n + 1) as u32 * HEX_NUM_OF_BITS_PER_DIGIT)) as u32 & HEX_DIGIT_MASK)
 }
 
 pub fn get_distance_between_unsigned<T>(a: T, b: T) -> T
